+++ packages/dynamodb-seeder/src/cli.ts
@@ -3,7 +3,7 @@
 import yargs from 'yargs'
 import { hideBin } from 'yargs/helpers'
 
-import validateCommand from './commands/validate'
+import { lintCommand, seedCommand } from './commands'
 
 const environment = {
   // eslint-disable-next-line @typescript-eslint/no-var-requires
@@ -15,7 +15,8 @@ yargs(hideBin(process.argv))
   .scriptName('dynamodb-seeder')
   .version('version', environment.version)
   .usage('$0 <cmd> [args]')
-  .command(validateCommand)
+  .command(lintCommand)
+  .command(seedCommand)
   .demandCommand()
   .help()
   .argv