+++ packages/dynamodb-seeder/src/commands/index.ts
@@ -1,0 +1,2 @@
+export { default as lintCommand } from './lint'
+export { default as seedCommand } from './seed'