+++ packages/dynamodb-seeder/src/commands/seed.test.ts
@@ -1,0 +1,5 @@
+describe('seed command', () => {
+  it('should pass', () => {
+    expect(true).toEqual(true)
+  })
+})