+++ packages/dynamodb-seeder/CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `seed` command
 
+### Changed
+- `validate` command renamed to `lint` (aliased to `validate`)
+
 ## [0.1.2] - 2022-05-22
 ### Added